aiServices/requirements.txt
# Flask API
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0

# Vision AI (ResNet50)
torch==2.9.0
torchvision==0.24.0
pillow==10.1.0

# Vertex AI (Gemini)
google-generativeai==0.8.5
python-dotenv==1.0.0
